+++ fixtures/002/helpers.js
@@ -1,0 +1,3 @@
+module.exports = {
+  upcase: str => str.toUpperCase(),
+};