+++ fixtures/002/input/Foo.sol
@@ -1,0 +1,6 @@
+// SPDX-License-Identifier: MIT
+
+pragma solidity ^0.6;
+
+contract Foo {
+}